opt(rx-dma-calypte): flip valid upon header buffer rollover
This eliminates the clearing of the valid bit in the host memory which
reduces the amount of writes to it. The valid value is initialized to 1
upon start of a channel and remains the same until a header pointer
rolls over which flips the value to 0. Upon next rollover, the value
flips to 1 and so on.