feat(cards-amd): enable Straddling on CQ/CC interfaces
The straddling has been enabled to keep the interface in concordance
with Intel devices that unilize this functionality by default and also
to improve throughput for DMA Calypte.

With regards of the updated Axi4SCompleter class, this only supports
assignment of signals that are on the Straddle-enabled interfaces but
does not utilize multiple regions. The _handle_cc_transaction needs to
be extended to be able to detect packets starting in the second region.

The _cq_reg does assign input transactions only to the first region.